v4: linux/386: run/pass more tests

+++ v4/lib/init.go
@@ -340,7 +340,7 @@ func (c *ctx) initializerStruct(w writer, n cc.Node, a []*cc.Initializer, t *cc.
 				}
 				f = in.Field()
 				sh := f.OffsetBits() + 8*int(f.Offset()-ogo)
-				b.w("(((%s)&%#0x)<<%d)", c.expr(w, in.AssignmentExpression, c.unsignedInts[gsz/8], exprDefault), uint(1)<<f.ValueBits()-1, sh)
+				b.w("(((%s)&%#0x)<<%d)", c.expr(w, in.AssignmentExpression, c.unsignedInts[gsz/8], exprDefault), uint64(1)<<f.ValueBits()-1, sh)
 			}
 			b.w(", ")
 			continue
@@ -589,7 +589,7 @@ func (c *ctx) initializerUnionOne(w writer, n cc.Node, a []*cc.Initializer, t *c
 	b.w("}{%sf: ", tag(preserve))
 	switch f := in.Field(); {
 	case f != nil && f.IsBitfield():
-		b.w("(((%s)&%#0x)<<%d)", c.expr(w, in.AssignmentExpression, c.unsignedInts[f.AccessBytes()], exprDefault), uint(1)<<f.ValueBits()-1, f.OffsetBits())
+		b.w("(((%s)&%#0x)<<%d)", c.expr(w, in.AssignmentExpression, c.unsignedInts[f.AccessBytes()], exprDefault), uint64(1)<<f.ValueBits()-1, f.OffsetBits())
 	default:
 		b.w("%s", c.expr(w, in.AssignmentExpression, in.Type(), exprDefault))
 	}
